requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
scipy>=1.11.0
scikit-image>=0.22.0
matplotlib>=3.7.0